requirements.txt
pinecone-haystack
haystack-ai
python-dotenv
pathlib
pypdf
sentence-transformers
python-multipart
trafilatura
streamlit
-e .